requirements.txt
streamlit==1.28.1
langchain==0.1.16
langchain-groq==0.1.3
groq==0.4.1
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.5.0
pandas==2.1.3
numpy>=1.24.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
streamlit-chat==0.1.1
plotly==5.17.0
ragas==0.0.22
datasets==2.14.6

